Made rule properties editable in the rule enabler dialog. This is done as an "override" on top of the rule properties configured for each rule in the ruleset XML files. Only existing properties can be edited; new ones can't be added, and existing ones can't be removed.
Implementation is slightly quick-n-dirty; I would prefer to
refactor PMD configuration to store rulesets defined in XML
files in the system filesystem, and allow import into that
set of rulesets from external files or the rulesets/* stuff
in the classpath. That would require a bit more surgery :)
and also some discussion to make sure my idea is not
significantly brain-damaged. So this will do for now.

Completes task 75427.


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@1951 51baf565-9d33-0410-a72c-fc3788e3496d